Session 9 - Particles completed Session 10 - Quaternions completed Created an EmptyScene for future usage Copied across my bezier curve class from my math lib(really need to create a proper library workflow for all this stuff) Created the QuatScene and ParticleScene

+++ Source/Profiler.h
@@ -82,7 +82,7 @@ class ProfileNode
 		_min = (time < _min) ? time : _min;
 		_max = (time > _max) ? time : _max;
 
-		if (_frameTimes.size() >= FRAMETIMES_MAXCOUNT)
+		if ((int)_frameTimes.size() >= FRAMETIMES_MAXCOUNT)
 			_frameTimes.pop();
 	}
